Refactoring code.
- Component file names should be capitalized.
- If there's a function that returns JSX, it's a component. Make it one.
- Use env vars for stuff that changes (API URLs for instance).
