+++ .gitignore
@@ -2,3 +2,4 @@
 /codespan-reporting/target
 a.out
 /letsencrypt
+.idea/
